chore(jans-cedarling): SparKV move tests to own file
